+++ src/main/java/com/verdantartifice/primalmagic/client/gui/grimoire/OtherIndexPage.java
@@ -4,10 +4,13 @@ import com.mojang.blaze3d.vertex.PoseStack;
 import com.verdantartifice.primalmagic.client.gui.GrimoireScreen;
 import com.verdantartifice.primalmagic.client.gui.widgets.grimoire.AttunementIndexButton;
 import com.verdantartifice.primalmagic.client.gui.widgets.grimoire.DisciplineButton;
+import com.verdantartifice.primalmagic.client.gui.widgets.grimoire.RecipeIndexButton;
 import com.verdantartifice.primalmagic.client.gui.widgets.grimoire.RuneEnchantmentIndexButton;
 import com.verdantartifice.primalmagic.client.gui.widgets.grimoire.StatisticsButton;
 import com.verdantartifice.primalmagic.common.research.ResearchDiscipline;
 import com.verdantartifice.primalmagic.common.research.ResearchDisciplines;
+import com.verdantartifice.primalmagic.common.research.ResearchEntries;
+import com.verdantartifice.primalmagic.common.research.ResearchEntry;
 import com.verdantartifice.primalmagic.common.research.ResearchManager;
 import com.verdantartifice.primalmagic.common.research.SimpleResearchKey;
 
@@ -44,6 +47,14 @@ public class OtherIndexPage extends AbstractPage {
         if (ResearchManager.isResearchComplete(mc.player, SimpleResearchKey.parse("ATTUNEMENTS"))) {
             text = new TranslatableComponent("primalmagic.grimoire.attunement_header");
             screen.addWidgetToScreen(new AttunementIndexButton(x + 12 + (side * 140), y, text, screen));
+            y += 12;
+        }
+        
+        // Add recipes button
+        ResearchEntry firstSteps = ResearchEntries.getEntry(SimpleResearchKey.FIRST_STEPS);
+        if (firstSteps != null && (firstSteps.isComplete(mc.player) || firstSteps.isInProgress(mc.player))) {
+            text = new TranslatableComponent("primalmagic.grimoire.recipe_index_header");
+            screen.addWidgetToScreen(new RecipeIndexButton(x + 12 + (side * 140), y, text, screen));
             y += 12;
         }